FIX: Change elapsed time representation from seconds to centiseconds for improved accuracy

+++ src/model/game/Game.java
@@ -213,7 +213,7 @@ public class Game extends GameBoard {
         return starCount;
     }
     // ELAPSED TIME //
-    public int getElapsedTime() {
+    public long getElapsedTime() {
         return _Timer.elapsedTime;
     }
